Should compile with modern VS. Working with VS2019
Progress to a working build.
Audio & Video (ie bink) are disabled
3DSMax 4 SDK is the biggest item blocking the build

+++ Generals/Code/GameEngine/Include/Common/AsciiString.h
@@ -366,7 +366,7 @@ inline AsciiString::~AsciiString()
 inline int AsciiString::getLength() const
 {
 	validate();
-	return m_data ? strlen(peek()) : 0;
+	return m_data ? (int)strlen(peek()) : 0;
 }
 
 // -----------------------------------------------------
