+++ Generals/Code/GameEngine/Include/Common/GameLOD.h
@@ -168,8 +168,8 @@ public:
 	Bool setDynamicLODLevel(DynamicGameLODLevel level);	///< set the current dynamic LOD level.
 	DynamicGameLODLevel getDynamicLODLevel(void) { return m_currentDynamicLOD;}
 	void init(void);	///<initialize tables of preset LOD's.
-	void setCurrentTextureReduction(Int val) {m_currentTextureReduction = val;}
-	Int getCurrentTextureReduction(void) {return m_currentTextureReduction;}
+	void setCurrentTextureReduction(Int val) {m_currentTextureReduction = (float)val;}
+	Int getCurrentTextureReduction(void) {return (int)m_currentTextureReduction;}
 	Int getStaticGameLODIndex(AsciiString name);
 	Int getDynamicGameLODIndex(AsciiString name);
 	inline Bool isParticleSkipped(void);
