+++ Generals/Code/GameEngine/Include/GameLogic/ObjectTypes.h
@@ -75,10 +75,10 @@ public:
 	Bool isInSet(const ThingTemplate* objectType) const;
 
 	// Is the set empty?
-	UnsignedInt getListSize(void) const { return m_objectTypes.size(); }
+	UnsignedInt getListSize(void) const { return (uint32_t)m_objectTypes.size(); }
 	
 	// I'd like to loop through, please.
-	AsciiString getNthInList( Int index ) const { return (index < getListSize()) ? m_objectTypes[index] : AsciiString::TheEmptyString; }
+	AsciiString getNthInList( Int index ) const { return (index < (int)getListSize()) ? m_objectTypes[index] : AsciiString::TheEmptyString; }
 
 	// Prep two arrays for usage with Player::countObjectsByThingTemplate
 	Int prepForPlayerCounting( std::vector<const ThingTemplate *>& templates, std::vector<Int>& counts);
