+++ Generals/Code/GameEngine/Source/GameLogic/System/RankInfo.cpp
@@ -76,7 +76,7 @@ void RankInfoStore::reset()
 //-----------------------------------------------------------------------------
 Int RankInfoStore::getRankLevelCount() const
 { 
-	return m_rankInfos.size(); 
+	return (int)m_rankInfos.size(); 
 }
 
 //-----------------------------------------------------------------------------
