+++ Generals/Code/GameEngine/Source/GameNetwork/GameSpy/Thread/ThreadUtils.cpp
@@ -32,7 +32,7 @@
 
 std::wstring MultiByteToWideCharSingleLine( const char *orig )
 {
-	Int len = strlen(orig);
+	Int len = (int)strlen(orig);
 	WideChar *dest = NEW WideChar[len+1];
 
 	MultiByteToWideChar(CP_UTF8, 0, orig, -1, dest, len);
@@ -65,7 +65,7 @@ std::wstring MultiByteToWideCharSingleLine( const char *orig )
 std::string WideCharStringToMultiByte( const WideChar *orig )
 {
 	std::string ret;
-	Int len = WideCharToMultiByte( CP_UTF8, 0, orig, wcslen(orig), NULL, 0, NULL, NULL ) + 1;
+	Int len = WideCharToMultiByte( CP_UTF8, 0, orig, (int)wcslen(orig), NULL, 0, NULL, NULL ) + 1;
 	if (len > 0)
 	{
 		char *dest = NEW char[len];
