+++ Generals/Code/Libraries/Source/WWVegas/WW3D2/vertmaterial.cpp
@@ -513,7 +513,7 @@ WW3DErrorType VertexMaterialClass::Load_W3D(ChunkLoadClass & cload)
 
 		char *extended_arg_buffer = MSGW3DNEWARRAY("VertexMaterialClassTemp") char[mapping0_arg_len + 10];
 		sprintf(extended_arg_buffer, "[Args]\n%s", mapping0_arg_buffer);
-		mapping0_arg_len = strlen(extended_arg_buffer) + 1;
+		mapping0_arg_len = (unsigned int)(strlen(extended_arg_buffer) + 1);
 
 		delete [] mapping0_arg_buffer;
 		mapping0_arg_buffer = NULL;
@@ -530,7 +530,7 @@ WW3DErrorType VertexMaterialClass::Load_W3D(ChunkLoadClass & cload)
 
 		char *extended_arg_buffer = MSGW3DNEWARRAY("VertexMaterialClassTemp") char[mapping1_arg_len + 20];
 		sprintf(extended_arg_buffer, "[Args]\n%s", mapping1_arg_buffer);
-		mapping1_arg_len = strlen(extended_arg_buffer) + 1;
+		mapping1_arg_len = (unsigned int)(strlen(extended_arg_buffer) + 1);
 
 		delete [] mapping1_arg_buffer;
 		mapping1_arg_buffer = NULL;
