+++ Generals/Code/Libraries/Source/WWVegas/WWLib/wwstring.h
@@ -210,7 +210,7 @@ StringClass::operator= (const TCHAR *string)
 {
 	if (string != 0) {
 
-		int len = _tcslen (string);
+		int len = (int)_tcslen (string);
 		Uninitialised_Grow (len+1);
 		Store_Length (len);
 
@@ -295,7 +295,7 @@ inline
 StringClass::StringClass (const TCHAR *string, bool hint_temporary)
 	:	m_Buffer (m_EmptyString)
 {
-	int len=string ? _tcsclen(string) : 0;
+	int len=string ? (int)_tcsclen(string) : 0;
 	if (hint_temporary || len>0) {
 		Get_String (len+1, hint_temporary);
 	}
@@ -459,7 +459,7 @@ StringClass::operator+= (const TCHAR *string)
 	WWASSERT (string != NULL);
 
 	int cur_len = Get_Length ();
-	int src_len = _tcslen (string);
+	int src_len = (int)_tcslen (string);
 	int new_len = cur_len + src_len;
 
 	//
@@ -629,7 +629,7 @@ StringClass::Get_Length (void) const
 		// we better manually get the string length.
 		//
 		if (length == 0) {
-			length = _tcslen (m_Buffer);
+			length = (int)_tcslen (m_Buffer);
 			((StringClass *)this)->Store_Length (length);
 		}
 	}
